Fix getHeaders.
A bug was introduced in retrieving headers from requests
during upgrading to GHC 8. The problem was that getHeaders
filtered key-value pairs by value, not by key, when collected
values associated with a key. This resulted an empty list of
values for virtually every key.

This fix includes a quickcheck test for getHeaders.