Remove page SubmissionList.
This is a usability improvement change. Assignments
were accessible via two distinct pages, and only one
of them is capable of submitting a solution, but the
other is used too frequently to read the assignment.

Assignments are accessible via page Submission only.
Functionality of SubmissionList (listing submissions) is
moved to the page Submission.

Also, an assignment is not accessible before start time.

Error messages on page Submission are made colorful using alerts.

Some typos are also corrected.