Work around Windows 10 1703 Edge vs Snap upload incompatiblity.
Previous versions of Edge are incompatible with Snap in the
sense that filenames in upload forms include double quotes
around the filename. This causes problems when checking
file extensions in upload handling.

Also, this patch relaxes the restriction that only those files
with .zip extension are accepted for uploading. The server
takes a look at the file signature (aka magic number) as a
fallback.