Remove page UserSubmissions. LTS 9.14. datetime(6) in MySQL.
List of submissions on UserSubmissions is available
on the page Evaluation.

Use higher resolution datetime(6) type in MySQL. This makes possible
a more deterministic sort of elements by time. For this,
persistent-mysql 2.6.2 needed, which implies an LTS version bump.

UI improvements:

Evaluation is improved: it shows a warning if the current
submission is not the latest.

SubmissionInfo is renamed to SubmissionState, and contains
no option such as no submission.

On the student section of the Home page links lead to the latest
submission of the student.

SubmissionState centralizes how submission states are shown, making
possible a more uniform representation across pages. Also, it
introduces colored badges.

Text areas may have different sizes per page, thanks to the type
Size in the module Bootstrap.

Performance:

Loading a student's assignments and submissions on the Home page
may be slightly faster, since now the code avoids some toList/fromList
conversions.