Dictionary files are compiled together with BE-AD.
Compilation at start-up is time-consuming because
each time dictionary files are recompiled unnecessarily.

Moreover, it also causes problems. BE-AD dumps core when
compiled without -O2 using GHC 8.0.1 for some reason.
Besides, it does not respect compiler flags of BE-AD, such
as dynamic-only compilation, and always tries the default
static compilation. In short, it causes more headaches then
it brings to the table.

Icons are removed from DictionaryFile, as it is not (and will not be)
used.