Submission, csv permission checks. UI improvements.
Submissions are not accessible to anyone via crafted url.
This holds for the owner for ballot box and isolated
submissions.

A csv of a group or course is only accessible to an admin
of that group or course, respectively.

Csv files are named after the group or the course.

Long submissions are not collapsed by default.