Drop combobox plugin. Improve group registration page.
Combobox plugin is dropped because of its clumsiness,
and the fact that going back in the browser history
erases contents of comboboxes.

On the group registration page, there is no default
group selected. This is to avoid accidental registration
on the default group.

Some labels are improved on the group registration page.