Optimize persistence functions related to the home page.
lastSubmission do not fetch all submissions of a user for an assignment
in order to select the last. It issues an 'order by' clause in the
SQL statement.

stateOfSubmission do not fetch all feedback all feedbacks of a submission
in order to determine the latest test feedback. It issues an 'order by' clause
and filters feedbacks by type.